requirements.txt
pytorch
pytorch-lightning
lightning[extra]

jupyter
